requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.4
APScheduler==3.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
decorator==5.1.1
Django==2.2.19
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
lxml==4.9.3
magic-filter==1.0.11
multidict==6.0.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
parse==1.20.1
pefile==2023.2.7
Pillow==9.3.0
pyaes==1.6.1
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.10
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.28.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.1
urllib3==1.26.14
validators==0.20.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
